+++ src/main/java/com/timeflowsystem/security/SecurityApplication.java
@@ -2,10 +2,13 @@ package com.timeflowsystem.security;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 
 @SpringBootApplication
 @EnableWebSecurity(debug = true)
+@EnableMethodSecurity(jsr250Enabled = true,securedEnabled = true) // By deault, the prePostEnabled is set to true,
+																  // that's why we don't need to set it here
 public class SecurityApplication {
 
 	public static void main(String[] args) {