+++ src/main/java/com/timeflowsystem/security/repository/UserAccountRepository.java
@@ -1,13 +1,13 @@
 package com.timeflowsystem.security.repository;
 
 import com.timeflowsystem.security.model.UserAccount;
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserAccountRepository extends CrudRepository<UserAccount,Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
 
     Optional<UserAccount> findByEmail(String email);